+++ src/main/java/com/monginis/ops/model/grngvn/GetGrnGvnDetails.java
@@ -21,7 +21,7 @@ public class GetGrnGvnDetails {
 
 	private float itemMrp;
 
-	private int grnGvnQty;
+	private float grnGvnQty;
 
 	private float grnGvnAmt;
 
@@ -61,7 +61,7 @@ public class GetGrnGvnDetails {
 
 	private int delStatus;
 
-	private int grnGvnQtyAuto;
+	private float grnGvnQtyAuto;
 
 	private String itemName;
 
@@ -99,9 +99,9 @@ public class GetGrnGvnDetails {
 
 	// 23 FEB new Fields to handle qty variation between entry(insert) and dispatch
 
-	int aprQtyGate;
-	int aprQtyStore;
-	int aprQtyAcc;
+	float aprQtyGate;
+	float aprQtyStore;
+	float aprQtyAcc;
 
 	float aprTaxableAmt;
 	float aprTotalTax;
@@ -178,11 +178,11 @@ public class GetGrnGvnDetails {
 		this.itemMrp = itemMrp;
 	}
 
-	public int getGrnGvnQty() {
+	public float getGrnGvnQty() {
 		return grnGvnQty;
 	}
 
-	public void setGrnGvnQty(int grnGvnQty) {
+	public void setGrnGvnQty(float grnGvnQty) {
 		this.grnGvnQty = grnGvnQty;
 	}
 
@@ -338,11 +338,11 @@ public class GetGrnGvnDetails {
 		this.delStatus = delStatus;
 	}
 
-	public int getGrnGvnQtyAuto() {
+	public float getGrnGvnQtyAuto() {
 		return grnGvnQtyAuto;
 	}
 
-	public void setGrnGvnQtyAuto(int grnGvnQtyAuto) {
+	public void setGrnGvnQtyAuto(float grnGvnQtyAuto) {
 		this.grnGvnQtyAuto = grnGvnQtyAuto;
 	}
 
@@ -474,27 +474,27 @@ public class GetGrnGvnDetails {
 		this.refInvoiceDate = refInvoiceDate;
 	}
 
-	public int getAprQtyGate() {
+	public float getAprQtyGate() {
 		return aprQtyGate;
 	}
 
-	public void setAprQtyGate(int aprQtyGate) {
+	public void setAprQtyGate(float aprQtyGate) {
 		this.aprQtyGate = aprQtyGate;
 	}
 
-	public int getAprQtyStore() {
+	public float getAprQtyStore() {
 		return aprQtyStore;
 	}
 
-	public void setAprQtyStore(int aprQtyStore) {
+	public void setAprQtyStore(float aprQtyStore) {
 		this.aprQtyStore = aprQtyStore;
 	}
 
-	public int getAprQtyAcc() {
+	public float getAprQtyAcc() {
 		return aprQtyAcc;
 	}
 
-	public void setAprQtyAcc(int aprQtyAcc) {
+	public void setAprQtyAcc(float aprQtyAcc) {
 		this.aprQtyAcc = aprQtyAcc;
 	}
